Reads xml data from lattes zip file

Reads xml data from lattes zip file

gld_read_zip(zip.in)

Arguments

zip.in
A (single) name of zip file containing a xml file

Value

A list with the following items:

tpesq

A dataframe with information about researchers

tpublic

A dataframe with information about publications

Examples

f.in <- system.file('extdata/3262699324398819.zip', package = 'GetLattesData')
my.l <- gld_read_zip(f.in)

#> 
#> Reading  3262699324398819.zip -  Marcelo Scherer Perlin
#> 	Found 21 published papers
#> 	Found 2 accepted paper(s)
#> 	Found 10 supervisions
#> 	Found 2 published books
#> 	Found 0 book chapters
#> 	Found 17 conference papers

my.l

#> $tpesq
#>                     name last.update       phd.institution phd.start.year
#> 1 Marcelo Scherer Perlin  2018-09-24 University of Reading           2007
#>   phd.end.year country.origin                major.field   minor.field
#> 1         2010         Brasil CIENCIAS_SOCIAIS_APLICADAS Administração
#>                id.file
#> 1 3262699324398819.zip
#> 
#> $tpublic.published
#>                 id.file                   name
#> 1  3262699324398819.zip Marcelo Scherer Perlin
#> 2  3262699324398819.zip Marcelo Scherer Perlin
#> 3  3262699324398819.zip Marcelo Scherer Perlin
#> 4  3262699324398819.zip Marcelo Scherer Perlin
#> 5  3262699324398819.zip Marcelo Scherer Perlin
#> 6  3262699324398819.zip Marcelo Scherer Perlin
#> 7  3262699324398819.zip Marcelo Scherer Perlin
#> 8  3262699324398819.zip Marcelo Scherer Perlin
#> 9  3262699324398819.zip Marcelo Scherer Perlin
#> 10 3262699324398819.zip Marcelo Scherer Perlin
#> 11 3262699324398819.zip Marcelo Scherer Perlin
#> 12 3262699324398819.zip Marcelo Scherer Perlin
#> 13 3262699324398819.zip Marcelo Scherer Perlin
#> 14 3262699324398819.zip Marcelo Scherer Perlin
#> 15 3262699324398819.zip Marcelo Scherer Perlin
#> 16 3262699324398819.zip Marcelo Scherer Perlin
#> 17 3262699324398819.zip Marcelo Scherer Perlin
#> 18 3262699324398819.zip Marcelo Scherer Perlin
#> 19 3262699324398819.zip Marcelo Scherer Perlin
#> 20 3262699324398819.zip Marcelo Scherer Perlin
#> 21 3262699324398819.zip Marcelo Scherer Perlin
#>                                                                                                                                                                          article.title
#> 1                                                                                                              Teoria do Caos aplicada aos Contratos de Café no Mercado de Derivativos
#> 2                                                                                                               Evaluation of Pairs Trading Strategy at the Brazilian Financial Market
#> 3  Non-Linear Modelling in the Brazilian Market: Evaluating the Forecasting Performance of NN Univariate Nearest Neighbor and SNN Simultaneous Nearest Neighbour Forecasting Algorithm
#> 4                                                                                                           Liquidity risks in the UK high-value payment system: an empirical analysis
#> 5                                                                                                      Os efeitos da introdução de agentes de liquidez no mercado acionário brasileiro
#> 6                                                                           The determinants of a cross market arbitrage opportunity: theory and evidence for the European bond market
#> 7                                                                                          Comunalidades na Liquidez: Evidências e Comportamento Intradiário para o Mercado Brasileiro
#> 8                                                                                                                                                  On the Performance of the Tick Test
#> 9                                                                                              A ESTRATÉGIA DE PARES NO MERCADO ACIONÁRIO BRASILEIRO: O IMPACTO DA FREQUÊNCIA DE DADOS
#> 10                                             Os pesquisadores, as publicações e os periódicos da área de Finanças no Brasil: Uma análise com base em currículos da plataforma Lattes
#> 11                                                                                                 Análise do Perfil dos Acadêmicos e de suas Publicações Científicas em Administração
#> 12                                                                         Análise de Integração Financeira entre o Mercado Acionário Brasileiro e o Argentino: Uma Abordagem Dinâmica
#> 13                                                                                                              Can We Predict the Financial Markets Based on Google's Search Queries?
#> 14                                                                                         The Brazilian scientific output published in journals: A study based on a large CV database
#> 15                                                                                                  THE FORECASTING POWER OF INTERNET SEARCH QUERIES IN THE BRAZILIAN FINANCIAL MARKET
#> 16                                                                                     GetHFData: A R package for downloading and aggregating high frequency trading data from Bovespa
#> 17                                                         A multistage stochastic programming asset-liability management model: an application to the Brazilian pension fund industry
#> 18                                                                                                               Portfolio management using realized covariances: Evidence from Brazil
#> 19                                                                                                                                         Mispricing in the odd lots market in Brazil
#> 20                                                                                                         Is predatory publishing a real threat? Evidence from a large database study
#> 21                                                                                         A Microestrutura do Tesoudo Direto: Sazonalidade do Fluxo de Ordens e a Formação de Spreads
#>    year  language
#> 1  2006 Português
#> 2  2009    Inglês
#> 3  2007    Inglês
#> 4  2011    Inglês
#> 5  2013 Português
#> 6  2013    Inglês
#> 7  2013 Português
#> 8  2013    Inglês
#> 9  2015    Inglês
#> 10 2015 Português
#> 11 2017 Português
#> 12 2016 Português
#> 13 2016    Inglês
#> 14 2017    Inglês
#> 15 2017    Inglês
#> 16 2016    Inglês
#> 17 2017    Inglês
#> 18 2017 Português
#> 19 2017 Português
#> 20 2018    Inglês
#> 21 2016 Português
#>                                                              journal.title
#> 1                       READ - Revista Eletrônica da Administração (UFRGS)
#> 2                                 Journal of Derivatives &#38; Hedge Funds
#> 3         REAd. Revista Eletrônica de Administração (Porto Alegre. Online)
#> 4                                     Bank of England Working Paper Series
#> 5                                Revista Brasileira de Finanças (Impresso)
#> 6                                                Annals of Finance (Print)
#> 7                                REVISTA BRASILEIRA DE FINANÇAS (IMPRESSO)
#> 8                            The Quarterly Review of Economics and Finance
#> 9                         RAM. Revista de Administração Mackenzie (Online)
#> 10                               Revista Brasileira de Finanças (Impresso)
#> 11                  RAC. Revista de Administração Contemporânea (Impresso)
#> 12 REVISTA BRASILEIRA DE FINANÇAS: RBFIN = RBFIN: BRAZILIAN FINANCE REVIEW
#> 13                                                  JOURNAL OF FORECASTING
#> 14                                                 Journal of Informetrics
#> 15                        RAM. REVISTA DE ADMINISTRAÇÃO MACKENZIE (ONLINE)
#> 16                               REVISTA BRASILEIRA DE FINANÇAS (IMPRESSO)
#> 17                                            OPTIMIZATION AND ENGINEERING
#> 18                                               Revista Economia da ANPEC
#> 19                         North American Journal of Economics and Finance
#> 20                                                          SCIENTOMETRICS
#> 21                                            Revista de Economia Aplicada
#>         ISSN order.aut n.authors
#> 1          -         1         2
#> 2  1753-9641         1         1
#> 3  1413-2311         1         2
#> 4  1749-9135         1         2
#> 5  1679-0731         1         1
#> 6  1614-2446         1         3
#> 7  1679-0731         2         3
#> 8  1062-9769         1         3
#> 9  1678-6971         2         2
#> 10 1679-0731         1         2
#> 11 1415-6555         2         3
#> 12 1984-5146         2         2
#> 13 0277-6693         1         4
#> 14 1751-1577         1         5
#> 15 1678-6971         3         3
#> 16 1679-0731         1         2
#> 17 1389-4420         3         5
#> 18 1517-7580         3         4
#> 19 1062-9408         2         3
#> 20 0138-9130         1         3
#> 21 1413-8050         1         1
#> 
#> $tpublic.accepted
#>                id.file                   name
#> 1 3262699324398819.zip Marcelo Scherer Perlin
#> 2 3262699324398819.zip Marcelo Scherer Perlin
#>                                                                     article.title
#> 1 A Consumer Credit Risk Structural Model based on Affordability: Balance at Risk
#> 2  Lotka&amp;#146;s law for the Brazilian scientific output published in journals
#>   year language                  journal.title      ISSN order.aut n.authors
#> 1 2019   Inglês         Journal of Credit Risk 1744-6619         1         3
#> 2 2018   Inglês JOURNAL OF INFORMATION SCIENCE 0165-5515         2         3
#> 
#> $tsupervisions
#>                 id.file                   name    situation type.course
#> 1  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA   ACADEMICO
#> 2  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA   ACADEMICO
#> 3  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA   ACADEMICO
#> 4  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA   ACADEMICO
#> 5  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA   ACADEMICO
#> 6  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA        <NA>
#> 7  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA            
#> 8  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA            
#> 9  3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA            
#> 10 3262699324398819.zip Marcelo Scherer Perlin    CONCLUIDA            
#> 11 3262699324398819.zip Marcelo Scherer Perlin EM ANDAMENTO   ACADEMICO
#> 12 3262699324398819.zip Marcelo Scherer Perlin EM ANDAMENTO        <NA>
#>                                      course                       std.name
#> 1                   Dissertação de mestrado Gladys Helena Albarracín Gómez
#> 2                   Dissertação de mestrado              Martin Pontuschka
#> 3                   Dissertação de mestrado           Henrique Pinto Ramos
#> 4                   Dissertação de mestrado                   Kadja Mendes
#> 5                   Dissertação de mestrado    Angélica Maria Lizarazo Roa
#> 6                         Tese de doutorado                 Mauro Mastella
#> 7  TRABALHO_DE_CONCLUSAO_DE_CURSO_GRADUACAO   Pedro Gonçalves Simões Pires
#> 8  TRABALHO_DE_CONCLUSAO_DE_CURSO_GRADUACAO     Thales Antoniolli Crestani
#> 9  TRABALHO_DE_CONCLUSAO_DE_CURSO_GRADUACAO    Nicolas Gil Rodrigues Nunes
#> 10 TRABALHO_DE_CONCLUSAO_DE_CURSO_GRADUACAO      André Luís Tovo Rodrigues
#> 11                  Dissertação de mestrado               Camila Bavaresco
#> 12                        Tese de doutorado           HENRIQUE PINTO RAMOS
#>    year.supervision
#> 1              2015
#> 2              2015
#> 3              2016
#> 4              2016
#> 5              2016
#> 6              2015
#> 7              2011
#> 8              2011
#> 9              2012
#> 10             2011
#> 11             2017
#> 12             2016
#> 
#> $tbooks
#>                id.file                   name
#> 1 3262699324398819.zip Marcelo Scherer Perlin
#> 2 3262699324398819.zip Marcelo Scherer Perlin
#>                                               book.title book.year
#> 1         Processing and Analyzing Financial Data with R      2017
#> 2 Processamento e Modelagem de Dados Financeiros com o R      2017
#>         book.type  book.lan     book.issn book.npages book.edition book.editor
#> 1 LIVRO_PUBLICADO Português    8592243556         398            1  Amazon-KDP
#> 2 LIVRO_PUBLICADO Português 9788592243517         223            1  Amazon-KDP
#> 
#> $tconferences
#>                 id.file                   name
#> 1  3262699324398819.zip Marcelo Scherer Perlin
#> 2  3262699324398819.zip Marcelo Scherer Perlin
#> 3  3262699324398819.zip Marcelo Scherer Perlin
#> 4  3262699324398819.zip Marcelo Scherer Perlin
#> 5  3262699324398819.zip Marcelo Scherer Perlin
#> 6  3262699324398819.zip Marcelo Scherer Perlin
#> 7  3262699324398819.zip Marcelo Scherer Perlin
#> 8  3262699324398819.zip Marcelo Scherer Perlin
#> 9  3262699324398819.zip Marcelo Scherer Perlin
#> 10 3262699324398819.zip Marcelo Scherer Perlin
#> 11 3262699324398819.zip Marcelo Scherer Perlin
#> 12 3262699324398819.zip Marcelo Scherer Perlin
#> 13 3262699324398819.zip Marcelo Scherer Perlin
#> 14 3262699324398819.zip Marcelo Scherer Perlin
#> 15 3262699324398819.zip Marcelo Scherer Perlin
#> 16 3262699324398819.zip Marcelo Scherer Perlin
#> 17 3262699324398819.zip Marcelo Scherer Perlin
#>                                                                                                                                               article.title
#> 1                                                                                   Teoria do Caos Aplicada aos Contratos de Café no mercado de Derivativos
#> 2                                                                        Redução das Necessidades de Caixa frente a Investimentos no Mercado de Derivativos
#> 3                                                                                                       O CAPM na Bolsa de São Paulo: Um Modelo Condicional
#> 4  Non Parametric Modelling in Major Latin America Market Indexes: An Analysis of the Performance from the Nearest Neighbor Algorithm in Trading Strategies
#> 5                                                                                    Evaluation of Pairs Trading Strategy at the Brazilian Financial Market
#> 6                                                The Determinants of a Cross Market Arbitrage Opportunity: Theory and Evidence for the European Bond Market
#> 7                                                     A Microstructure Model for Spillover Effects in Price Discovery: A Study for the European Bond Market
#> 8                                                                                                                       On the Performance of the Tick Test
#> 9                                                                                                      Estimating the Intensity of News Based on Trade Data
#> 10                                                          A Microestrutura do Tesouro Direto: Sazonalidade da Demanda e o Processo de Formação de Spreads
#> 11                                                                                   Can we predict the financial markets based on Google's search queries?
#> 12             Os pesquisadores, as publica ções e os periódicos da area de Finan ças no Brasil: Uma anáalise com base em curr í culos da plataforma Lattes
#> 13                                                                                   Can We Predict the Financial Markets Based on Google's Search Queries?
#> 14                                                                                                                                    BaR - Balance at Risk
#> 15                                                          GetHFData: A R package for downloading and aggregating high frequency trading data from Bovespa
#> 16                                                                               Creating and Grading Latex Exams with Randomized Content using RndTexExams
#> 17                                                                                                                              CAPM e o Mercado Brasileiro
#>    article.year event.classification
#> 1          2004             NACIONAL
#> 2          2004             NACIONAL
#> 3          2004             NACIONAL
#> 4          2006        INTERNACIONAL
#> 5          2007             NACIONAL
#> 6          2011             NACIONAL
#> 7          2011             NACIONAL
#> 8          2012        INTERNACIONAL
#> 9          2013        INTERNACIONAL
#> 10         2013             NACIONAL
#> 11         2015             NACIONAL
#> 12         2015             NACIONAL
#> 13         2015        NAO_INFORMADO
#> 14         2016             NACIONAL
#> 15         2017        INTERNACIONAL
#> 16         2017        INTERNACIONAL
#> 17         2004             NACIONAL
#>                                                    event.name event.isbn
#> 1                      Quarto Encontro Brasileiro de Finanças           
#> 2            XXIV Encontro Nacional de Engenharia de Produção           
#> 3           4º Congresso USP de Controladoria e Contabilidade           
#> 4      BALAS - Business Associacion for Latin America Studies           
#> 5                          7° Encontro Brasileiro de Finanças           
#> 6                         11º Encontro Brasileiro em Finanças           
#> 7                         11º Encontro Brasileiro em Finanças           
#> 8    European Financial Management Association Annual Meeting           
#> 9  European Financial Management Association - Annual Meeting           
#> 10                            Encontro Brasileiro de Finanças           
#> 11                        15º Encontro Brasileiro de Finanças           
#> 12                        15º Encontro Brasileiro de Finanças           
#> 13                                                        FMA           
#> 14                        XVI Encontra Brasileiro de Finanças           
#> 15                                               R in Finance           
#> 16            II Seminário Internacional de Estatística com R           
#> 17          4º Congresso USP de Controladoria e Contabilidade           
#>            event.city
#> 1      Rio de Janeiro
#> 2  Florianópolis - SC
#> 3      São Paulo - SP
#> 4           Lima/Peru
#> 5       São Paulo, SP
#> 6      Rio de Janeiro
#> 7      Rio de Janeiro
#> 8  Barcelona, Espanha
#> 9          Reading/UK
#> 10     Rio de Janeiro
#> 11     São Paulo - SP
#> 12     São Paulo - SP
#> 13            Orlando
#> 14     Rio de Janeiro
#> 15         Chicago/IL
#> 16         Niteroi/RJ
#> 17     São Paulo - SP
#>